+++ PerseusMeteo/FirstPartyCode/Helpers/MessageLabel.swift
@@ -50,9 +50,9 @@ public typealias MyCustomLabel = UILabel
 public typealias MyCustomLabel = NSTextField
 #endif
 
-public let DEFAULT_MESSAGE_LABEL_TEXT = "Have a great time."
+public let DEFAULT_MESSAGE_LABEL_TEXT = "Have a great time"
 
-public class MessageLabel: MyCustomLabel {
+public class MessageLabel: MyCustomLabel, PerseusDelegatedMessage {
 
     private var messageDeepCounter: Int = 0