+++ PerseusMeteo/TheCrownOfStars/SupportingStars/PDMSupportingStar.swift
@@ -1,8 +1,8 @@
 //
 //  PDMSupportingStar.swift
-//  Version: 2.0.3
+//  Version: 2.1.0
 //
-//  The Darkness Support Classes (PerseusUISystemKit previously)
+//  The Darkness Support (PerseusUISystemKit previously)
 //
 //
 //  For iOS and macOS only. Use Stars to adopt for the specifics you need.
@@ -42,6 +42,8 @@
 //
 // swiftlint:disable file_length
 //
+
+// import PerseusDarkMode
 
 #if canImport(UIKit)
 import UIKit
@@ -561,7 +563,7 @@ public class DarkModeImageView: UIImageView {
         }
     }
 
-    private(set) var darkModeObserver: DarkModeObserver?
+    private(set) var theDarknessTrigger: DarkModeObserver?
 
     private(set) var light: UIImage?
     private(set) var dark: UIImage?
@@ -577,7 +579,7 @@ public class DarkModeImageView: UIImageView {
     }
 
     private func configure() {
-        darkModeObserver = DarkModeObserver { style in
+        theDarknessTrigger = DarkModeObserver { style in
             self.image = style == .light ? self.light : self.dark
         }
 
@@ -588,7 +590,7 @@ public class DarkModeImageView: UIImageView {
         self.light = light
         self.dark = dark
 
-        darkModeObserver?.action = { style in
+        theDarknessTrigger?.action = { style in
             self.image = style == .light ? self.light : self.dark
         }