+++ SnowmanTests/UnitTests/TranslationTests.swift
@@ -30,8 +30,8 @@ class TranslationTests: XCTestCase {
 
         // assert
 
-        XCTAssertEqual("Have a great time.".localizedExpectation,
-                       "Have a great time.".localizedValue)
+        XCTAssertEqual("Have a great time".localizedExpectation,
+                       "Have a great time".localizedValue)
 
         XCTAssertEqual("Greetings".localizedExpectation,
                        "Greetings".localizedValue)
@@ -72,6 +72,15 @@ class TranslationTests: XCTestCase {
 
         XCTAssertEqual("Default: N/A".localizedExpectation,
                        "Default: N/A".localizedValue)
+
+        XCTAssertEqual("P2P".localizedExpectation,
+                       "P2P".localizedValue)
+
+        XCTAssertEqual("P2P stands for Person to Person".localizedExpectation,
+                       "P2P stands for Person to Person".localizedValue)
+
+        XCTAssertEqual("API key is either rejected or empty".localizedExpectation,
+                       "API key is either rejected or empty".localizedValue)
     }
 
     // MARK: - End-User Messages
@@ -103,9 +112,6 @@ class TranslationTests: XCTestCase {
 
         XCTAssertEqual("Current Location should be selected".localizedExpectation,
                        "Current Location should be selected".localizedValue)
-
-        XCTAssertEqual("Favorites limited to".localizedExpectation,
-                       "Favorites limited to".localizedValue)
 
         XCTAssertEqual("Current neither to add nor to remove!".localizedExpectation,
                        "Current neither to add nor to remove!".localizedValue)
@@ -515,6 +521,10 @@ class TranslationTests: XCTestCase {
                        "Button: License".localizedValue)
         XCTAssertEqual("Button: Terms & Conditions".localizedExpectation,
                        "Button: Terms & Conditions".localizedValue)
+        XCTAssertEqual("Tab: Essentials".localizedExpectation,
+                       "Tab: Essentials".localizedValue)
+        XCTAssertEqual("Tab: Log".localizedExpectation,
+                       "Tab: Log".localizedValue)
     }
 
     // MARK: - Weather-conditions descriptions